+++ include/openrave/robot.h
@@ -30,15 +30,41 @@ namespace OpenRAVE {
 class OPENRAVE_API RobotBase : public KinBody
 {
 public:
+    /// \brief holds all user-set manipulator information used to initialize the Manipulator class
+    class OPENRAVE_API ManipulatorInfo
+    {
+public:
+        ManipulatorInfo() : _vdirection(0,0,1) {
+        }
+        virtual ~ManipulatorInfo() {
+        }
+
+        std::string _name;
+        std::string _sBaseLinkName, _sEffectorLinkName; ///< name of the base and effector links of the robot used to determine the chain
+        Transform _tLocalTool;
+        std::vector<dReal> _vClosingDirection; ///< the normal direction to move joints to 'close' the hand
+        Vector _vdirection;
+        std::string _sIkSolverXMLId; ///< xml id of the IkSolver interface to attach
+        std::vector<std::string> _vGripperJointNames;         ///< names of the gripper joints
+    };
+    typedef boost::shared_ptr<ManipulatorInfo> ManipulatorInfoPtr;
+
     /// \brief Defines a chain of joints for an arm and set of joints for a gripper. Simplifies operating with them.
     class OPENRAVE_API Manipulator : public boost::enable_shared_from_this<Manipulator>
     {
-        Manipulator(RobotBasePtr probot);
+        Manipulator(RobotBasePtr probot, const ManipulatorInfo& info);
         Manipulator(const Manipulator &r);
-        Manipulator(RobotBasePtr probot, const Manipulator &r);
+
+        /// \brief can switch the underyling robot
+        Manipulator(RobotBasePtr probot, boost::shared_ptr<Manipulator const> r);
 
 public:
         virtual ~Manipulator();
+
+        /// \brief return a serializable info holding everything to initialize a manipulator
+        inline const ManipulatorInfo& GetInfo() const {
+            return _info;
+        }
 
         /// \brief Return the transformation of the manipulator frame
         ///
@@ -54,37 +80,37 @@ public:
         }
 
         virtual const std::string& GetName() const {
-            return _name;
+            return _info._name;
         }
         virtual RobotBasePtr GetRobot() const {
-            return RobotBasePtr(_probot);
+            return RobotBasePtr(__probot);
         }
 
         /// \brief Sets the ik solver and initializes it with the current manipulator.
         ///
         /// Due to complications with translation,rotation,direction,and ray ik,
-        /// the ik solver should take into account the grasp transform (_tLocalTool) internally.
+        /// the ik solver should take into account the grasp transform (_info._tLocalTool) internally.
         /// The actual ik primitives are transformed into the base frame only.
         virtual bool SetIkSolver(IkSolverBasePtr iksolver);
 
         /// \brief Returns the currently set ik solver
         virtual IkSolverBasePtr GetIkSolver() const {
-            return _pIkSolver;
+            return __pIkSolver;
         }
 
         /// \brief the base used for the iksolver
         virtual LinkPtr GetBase() const {
-            return _pBase;
+            return __pBase;
         }
 
         /// \brief the end effector link (used to define workspace distance)
         virtual LinkPtr GetEndEffector() const {
-            return _pEndEffector;
+            return __pEffector;
         }
 
         /// \brief Return transform with respect to end effector defining the grasp coordinate system
         virtual Transform GetLocalToolTransform() const {
-            return _tLocalTool;
+            return _info._tLocalTool;
         }
 
         /// \brief Sets the local tool transform with respect to the end effector.
@@ -116,12 +142,12 @@ public:
 
         /// \brief return the normal direction to move joints to 'close' the hand
         virtual const std::vector<dReal>& GetClosingDirection() const {
-            return _vClosingDirection;
+            return _info._vClosingDirection;
         }
 
         /// \brief direction of palm/head/manipulator used for approaching. defined inside the manipulator/grasp coordinate system
         virtual Vector GetLocalToolDirection() const {
-            return _vdirection;
+            return _info._vdirection;
         }
 
         /// \deprecated (11/10/15) use GetLocalToolDirection
@@ -274,19 +300,16 @@ public:
         /// This includes joint axes, joint positions, and final grasp transform. Hash is used to cache the solvers.
         virtual const std::string& GetKinematicsStructureHash() const;
 protected:
-        std::string _name;
-        LinkPtr _pBase, _pEndEffector;
-        Transform _tLocalTool;
-        std::vector<dReal> _vClosingDirection;
-        Vector _vdirection;
-        IkSolverBasePtr _pIkSolver;
-        std::string _strIkSolver;
-        std::vector<std::string> _vgripperjointnames;         ///< names of the gripper joints
+        /// \brief compute internal information from user-set info
+        virtual void _ComputeInternalInformation();
 
+        ManipulatorInfo _info; ///< user-set information
 private:
-        RobotBaseWeakPtr _probot;
+        RobotBaseWeakPtr __probot;
+        LinkPtr __pBase, __pEffector; ///< contains weak links to robot
         std::vector<int> __vgripperdofindices, __varmdofindices;
         ConfigurationSpecification __armspec; ///< reflects __varmdofindices
+        IkSolverBasePtr __pIkSolver;
         mutable std::string __hashstructure, __hashkinematicsstructure;
 
 #ifdef RAVE_PRIVATE
@@ -590,7 +613,7 @@ private:
     /// Will change the robot structure hash..
     /// \return the new manipulator attached to the robot
     /// \throw openrave_exception If there exists a manipulator with the same name, will throw an exception
-    virtual ManipulatorPtr AddManipulator(ManipulatorConstPtr manip);
+    virtual ManipulatorPtr AddManipulator(const ManipulatorInfo& manipinfo);
 
     /// \brief removes a manipulator from the robot list.
     ///